dockerfile: update mysql to 10.3.13
This includes a lot of changes. To be honest, I'm not sure if this
will be fully compatible, but if you were really migrating between major
mysql versions you'd have to handle your own upgrade regardless.

File a bug if you find specific things to improve in this story.

Other changes:
- upstream (Alpine port) did a lot of changes. Adapt.
- upstream (mariadb) changed user table format. Adapt.
- a lot of changes to first time init; install and bootstrap
- removed a few more files
- Readme updates