Refactor root password generation
With MariaDB 10.14.x a lot of internals changed, most notably
(for us) the `user` table becoming a view, meaning we cannot
insert to it.

We fix this by moving the host line to image creation and
the optional password creation as a regular sql command as part
of the first init.