Create workflows for github actions
Introduce a build and test workflow. Both actually builds it - caching
will be a future optimization.

Also, introdue `bin/build-image.sh` which will be the central point
for building the container images; be it CI or local.

This will fix #4 since we now know if we break things between updates.

Fixes: https://github.com/jbergstroem/mariadb-alpine/issues/4