dockerfile: optimize cleanup
We previously called `tail` on output from `apk`. This is a remnant from
an older script that handled output differently.

Also, remove all pam functionality since we don't use it.

This means we now remove more files that shouldn't be part of the distribution.
For this to work as previously, we need to manually create `/etc/my.cnf`.